Added in Trees, Ore Spawning, Survival Inventory, Block Durability, Dirt changing to grass if no block is above it, Grass turning to dirt if block is above it,  Better Terrain Generation, Created and Started a LivingEntity class, Added in solid and non-solid blocks.

+++ src/com/frostbyte/util/MathUtil.java
@@ -4,7 +4,7 @@ import com.frostbyte.display.Location;
 import com.frostbyte.display.Vector;
 
 public class MathUtil {
-	public static boolean getRandomBoolean(int amount) {
+	public static boolean getRandomBoolean(double amount) {
 		return Math.random() < (amount / 100.0);
 	}
